Add SiteKit and WordPress modules.
